cloud_function/requirements.txt
functions-framework==3.*
google-cloud-logging>=3.11
requests==2.*
flask==3.*
google-cloud-bigquery==3.*
